move data collection into a module
